Update option_scraperBS.py
I removed a section of the code to see if it could run differently. The reason I removed this function was that it pulled from a website that I never got any information from. In turn it seemed useless, but it broke my code completely so I put it back